// src/app/models/group.ts
export class Group {
    live_group_id : number;
    owner_id: number;
    live_group_name: string;
    live_group_description: string;
    live_group_picture: string;
    live_group_status: any;
    owner_name: string;
    picture: string;
    total_members: any;
    is_owner: any;
    date : string;
}